Extract scan handling into ScanManager
