+++ android/app/src/main/java/org/bitcoinppl/cove/AppManager.kt
@@ -296,245 +296,6 @@ class AppManager private constructor() : FfiReconcile {
         sheetState = TaggedItem(AppSheetState.Nfc)
     }
 
-    /**
-     * Handle scanned QR code data by parsing and routing based on content type
-     * Matches iOS implementation in CoveApp.swift
-     */
-    fun handleMultiFormat(multiFormat: MultiFormat) {
-        try {
-            when (multiFormat) {
-                is MultiFormat.Mnemonic -> {
-                    multiFormat.v1.use { mnemonic ->
-                        importHotWallet(mnemonic.words())
-                    }
-                }
-
-                is MultiFormat.HardwareExport -> {
-                    importColdWallet(multiFormat.v1)
-                }
-
-                is MultiFormat.Address -> {
-                    handleAddress(multiFormat.v1)
-                }
-
-                is MultiFormat.Transaction -> {
-                    handleTransaction(multiFormat.v1)
-                }
-
-                is MultiFormat.SignedPsbt -> {
-                    handleSignedPsbt(multiFormat.v1)
-                }
-
-                is MultiFormat.TapSignerUnused -> {
-                    alertState = TaggedItem(AppAlertState.UninitializedTapSigner(multiFormat.v1))
-                }
-
-                is MultiFormat.TapSignerReady -> {
-                    val wallet = findTapSignerWallet(multiFormat.v1)
-                    if (wallet != null) {
-                        alertState = TaggedItem(AppAlertState.TapSignerWalletFound(wallet.id))
-                    } else {
-                        alertState = TaggedItem(AppAlertState.InitializedTapSigner(multiFormat.v1))
-                    }
-                }
-
-                is MultiFormat.Bip329Labels -> {
-                    val selectedWallet = database.globalConfig().selectedWallet()
-                    if (selectedWallet == null) {
-                        alertState =
-                            TaggedItem(
-                                AppAlertState.InvalidFileFormat(
-                                    "Currently BIP329 labels must be imported through the wallet actions",
-                                ),
-                            )
-                        return
-                    }
-
-                    // import the labels
-                    try {
-                        LabelManager(id = selectedWallet).use { it.importLabels(multiFormat.v1) }
-                        alertState = TaggedItem(AppAlertState.ImportedLabelsSuccessfully)
-
-                        // when labels are imported, refresh transactions with updated labels
-                        walletManager?.let { wm ->
-                            mainScope.launch {
-                                wm.rust.getTransactions()
-                            }
-                        }
-                    } catch (e: Exception) {
-                        Log.e(tag, "Failed to import labels", e)
-                        alertState =
-                            TaggedItem(
-                                AppAlertState.InvalidFileFormat(
-                                    e.message ?: "Failed to import labels",
-                                ),
-                            )
-                    }
-                }
-            }
-        } catch (e: Exception) {
-            Log.e(tag, "Unable to handle scanned code", e)
-            alertState =
-                TaggedItem(
-                    AppAlertState.InvalidFileFormat(e.message ?: "Unknown error"),
-                )
-        }
-    }
-
-    /**
-     * Import hot wallet from mnemonic words
-     */
-    private fun importHotWallet(words: List<String>) {
-        val manager = ImportWalletManager()
-        try {
-            val walletMetadata = manager.rust.importWallet(listOf(words))
-            rust.selectWallet(walletMetadata.id)
-        } catch (e: ImportWalletException.InvalidWordGroup) {
-            Log.d(tag, "Invalid word group detected")
-            alertState = TaggedItem(AppAlertState.InvalidWordGroup)
-        } catch (e: ImportWalletException.WalletAlreadyExists) {
-            Log.w(tag, "Attempted to import words for an existing hot wallet: ${e.v1}")
-            alertState = TaggedItem(AppAlertState.DuplicateWallet(e.v1))
-            try {
-                rust.selectWallet(e.v1)
-            } catch (selectError: Exception) {
-                Log.e(tag, "Unable to select existing wallet", selectError)
-            }
-        } catch (e: Exception) {
-            Log.e(tag, "Unable to import wallet", e)
-            alertState =
-                TaggedItem(
-                    AppAlertState.ErrorImportingHotWallet(e.message ?: "Unknown error"),
-                )
-        } finally {
-            manager.close()
-        }
-    }
-
-    /**
-     * Import cold wallet from hardware export
-     */
-    private fun importColdWallet(export: HardwareExport) {
-        try {
-            val wallet = Wallet.newFromExport(export)
-            try {
-                val id = wallet.id()
-                Log.d(tag, "Imported Wallet: $id")
-                alertState = TaggedItem(AppAlertState.ImportedSuccessfully)
-
-                // if we're not already on this wallet, navigate to it
-                if (walletManager?.id != id) {
-                    rust.selectWallet(id)
-                }
-
-                // upgrade watch-only → cold in-place
-                if (walletManager?.id == id && walletManager?.walletMetadata?.walletType != WalletType.HOT) {
-                    try {
-                        walletManager?.rust?.setWalletType(WalletType.COLD)
-                    } catch (e: Exception) {
-                        Log.e(tag, "Failed to set wallet type to cold", e)
-                    }
-                }
-            } finally {
-                wallet.close()
-            }
-        } catch (e: WalletException.WalletAlreadyExists) {
-            alertState = TaggedItem(AppAlertState.DuplicateWallet(e.v1))
-            try {
-                rust.selectWallet(e.v1)
-            } catch (selectError: Exception) {
-                Log.e(tag, "Unable to select existing wallet", selectError)
-                alertState = TaggedItem(AppAlertState.UnableToSelectWallet)
-            }
-        } catch (e: Exception) {
-            Log.e(tag, "Error importing hardware wallet", e)
-            alertState =
-                TaggedItem(
-                    AppAlertState.ErrorImportingHardwareWallet(e.message ?: "Unknown error"),
-                )
-        }
-    }
-
-    /**
-     * Handle scanned bitcoin address
-     */
-    private fun handleAddress(addressWithNetwork: AddressWithNetwork) {
-        val currentNetwork = database.globalConfig().selectedNetwork()
-        val address = addressWithNetwork.address()
-        val network = addressWithNetwork.network()
-        val selectedWallet = database.globalConfig().selectedWallet()
-
-        if (selectedWallet == null) {
-            alertState = TaggedItem(AppAlertState.NoWalletSelected(address))
-            return
-        }
-
-        if (!addressWithNetwork.isValidForNetwork(currentNetwork)) {
-            alertState =
-                TaggedItem(
-                    AppAlertState.AddressWrongNetwork(
-                        address = address,
-                        network = network,
-                        currentNetwork = currentNetwork,
-                    ),
-                )
-            return
-        }
-
-        val amount = addressWithNetwork.amount()
-        alertState = TaggedItem(AppAlertState.FoundAddress(address, amount))
-    }
-
-    /**
-     * Handle scanned signed transaction
-     */
-    private fun handleTransaction(transaction: BitcoinTransaction) {
-        Log.d(tag, "Received BitcoinTransaction: $transaction: ${transaction.txIdHash()}")
-
-        val db = database.unsignedTransactions()
-        val txnRecord = db.getTx(transaction.txId())
-
-        if (txnRecord == null) {
-            Log.e(tag, "No unsigned transaction found for ${transaction.txId()}")
-            alertState = TaggedItem(AppAlertState.NoUnsignedTransactionFound(transaction.txId()))
-            return
-        }
-
-        val route =
-            RouteFactory().sendConfirm(
-                id = txnRecord.walletId(),
-                details = txnRecord.confirmDetails(),
-                signedTransaction = transaction,
-            )
-
-        pushRoute(route)
-    }
-
-    /**
-     * Handle scanned signed PSBT
-     */
-    private fun handleSignedPsbt(psbt: Psbt) {
-        Log.d(tag, "Received signed PSBT: ${psbt.txId()}")
-
-        val db = database.unsignedTransactions()
-        val txnRecord = db.getTx(psbt.txId())
-
-        if (txnRecord == null) {
-            Log.e(tag, "No unsigned transaction found for PSBT ${psbt.txId()}")
-            alertState = TaggedItem(AppAlertState.NoUnsignedTransactionFound(psbt.txId()))
-            return
-        }
-
-        val route =
-            RouteFactory().sendConfirm(
-                id = txnRecord.walletId(),
-                details = txnRecord.confirmDetails(),
-                signedPsbt = psbt,
-            )
-
-        pushRoute(route)
-    }
-
     fun resetRoute(to: List<Route>) {
         if (to.size > 1) {
             rust.resetNestedRoutesTo(to[0], to.drop(1))
